obra-ya/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
anthropic>=0.40.0
sqlalchemy>=2.0.35
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.10.0
python-multipart>=0.0.9
openai>=1.50.0
stripe>=7.0.0
twilio>=9.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
